+++ CHANGELOG.md
@@ -6,6 +6,9 @@ v0.4.0 is a limited availability feature release. It is supported for all usage.
 
 1. Fixes an issue where headers were not passed correctly to the `eachBatch` callback (#130).
 2. Add support for an Admin API to list a consumer group's offsets (#49).
+3. Reduce consumer poll timeout to nil and add wakeups for new messages. This improves
+   the consumer efficiency, and resolves issues while running multiple consumers within
+   the same node process (#135).
 
 
 # confluent-kafka-javascript v0.3.0